Fix share analysis scripts

+++ scripts/treeCreation/analyzeFrequency.sh
@@ -2,16 +2,30 @@ function getSum {
   awk '{sum += $1; square += $1^2} END {print sqrt(square / NR - (sum/NR)^2)" "sum/NR" "NR}'
 }
 
+treeFolder=$1
+if [ ! -d $treeFolder ]
+then
+	echo "TreeFolder was no folder: $treeFolder"
+	exit 1
+fi
+simpleTraceFolder=$treeFolder/simpleTraces
+if [ ! -d $simpleTraceFolder ]
+then
+	echo "SimpleTraceFolder was not existing: $simpleTraceFolder"
+	exit 1
+fi
+
 rm invocations.txt
 
 for line in $(cat ../regressions.csv)
 do
-	benchmark=$(echo $line | awk -F';' '{print $2}')
+	regression=$(echo $line | awk -F';' '{print $1}')
+	selectedBenchmark=$(echo $line | awk -F';' '{print $2}')
 	changedMethod=$(echo $line | awk -F';' '{print $3}')
-	changedMethodCalls=$(cat $1/simpleTraces/root-$selectedBenchmark_* | grep  $changedMethod | wc -l)
-	calls=$(cat $1/simpleTraces/root-$selectedBenchmark_* | wc -l)
-	echo "ChangedCalls: $changedMethodCalls OverallCalls: $calls"
+	changedMethodCalls=$(cat $simpleTraceFolder/root-"$selectedBenchmark"* | grep  $changedMethod | wc -l)
+	calls=$(cat $1/simpleTraces/root-"$selectedBenchmark"* | wc -l)
+	echo "$regression ChangedCalls: $changedMethodCalls OverallCalls: $calls"
 done > invocations.txt
 
 echo -n "Overall share of changed method: "
-cat invocations.txt | awk '{print $2/$4}' | getSum
+cat invocations.txt | awk '{if ($5 != 0) {print $3/$5}}' | getSum
